# docs/API_OPENAPI.yaml
openapi: 3.0.3
info:
  title: Course Companion FTE API
  description: |
    **Zero-Backend-LLM Educational Platform API**

    A Digital Full-Time Equivalent (FTE) educational tutor that serves thousands of students simultaneously.

    ## Architecture

    - **Phase 1 (Zero-LLM):** Deterministic content delivery, rule-based quizzes, progress tracking
    - **Phase 2 (Hybrid):** Premium AI features (LLM grading, AI mentor, adaptive learning)
    - **Phase 3 (Web):** Full web application with all features

    ## Authentication

    Most endpoints require JWT authentication via `Authorization: Bearer <token>` header.

    ## Rate Limits

    - Free users: 100 requests/hour
    - Premium users: 1000 requests/hour
    - PRO users: Unlimited

  version: 1.0.0
  contact:
    name: Course Companion FTE Team
    url: https://github.com/course-companion-fte
  license:
    name: MIT
servers:
  - url: http://92.113.147.250:3505
    description: Production server
  - url: http://localhost:3505
    description: Local development server
...
paths:
  /api/v3/tutor/content/chapters:
    get:
      tags: [Content]
      summary: List all chapters
      description: |
        Returns all course chapters with metadata and completion status.

        **Phase:** Zero-LLM (deterministic)
      operationId: listChapters
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of chapters retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  chapters:
                    type: array
                    items:
                      $ref: '#/components/schemas/ChapterListItem'
                  total:
                    type: integer
                  completion_percentage:
                    type: number
  /api/v3/tutor/content/chapters/{chapter_id}:
    get:
      tags: [Content]
      summary: Get chapter content
      description: |
        Returns full chapter content with completion status.

        **Phase:** Zero-LLM (serves content verbatim)
      operationId: getChapter
      security:
        - BearerAuth: []
      parameters:
        - name: chapter_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Chapter content retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterContent'
        '404':
          description: Chapter not found
  /api/v3/tutor/content/chapters/{chapter_id}/navigation:
    get:
      tags: [Content]
      summary: Get chapter navigation
      description: Returns previous and next chapter for navigation
      operationId: getNavigation
      security:
        - BearerAuth: []
      parameters:
        - name: chapter_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Navigation context retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChapterNavigation'
  /api/v3/tutor/content/search:
    get:
      tags: [Content]
      summary: Search course content
      description: Keyword search across all course content
      operationId: searchContent
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          description: Search query
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  query:
                    type: string
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/SearchResult'
                  total:
                    type: integer
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ChapterListItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        order:
          type: integer
        difficulty_level:
          type: string
          enum: [BEGINNER, INTERMEDIATE, ADVANCED]
        estimated_time:
          type: integer
          description: Estimated reading time in minutes
        completed:
          type: boolean
        quiz_id:
          type: string
          format: uuid
          nullable: true
    ChapterContent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        order:
          type: integer
        difficulty_level:
          type: string
        estimated_time:
          type: integer
        content:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        quiz_id:
          type: string
          format: uuid
          nullable: true
        completed:
          type: boolean
    ChapterNavigation:
      type: object
      properties:
        current:
          $ref: '#/components/schemas/ChapterListItem'
        previous:
          $ref: '#/components/schemas/ChapterListItem'
        next:
          $ref: '#/components/schemas/ChapterListItem'
        can_continue:
          type: boolean
    SearchResult:
      type: object
      properties:
        chapter_id:
          type: string
          format: uuid
        title:
          type: string
        snippet:
          type: string
        relevance_score:
          type: number
...